Pulled HDProcessor class out of HDVizProcess command line.
The core functionality of the HDVizProcess executable is now pulled
into a library. The structure of the processing is kept the same in
this change. The CLI HDVizProcess tool now longer needs to know
directly about the inner workings of the processing, and instead
now only deals with command lin argument processing and exception
handling.