Refactor of getExtremaWidths method for HDVizData types.
The original HDVizData object (now FileCachedHDVizDataImpl) was
introducing a scaling operation to ExtremaWidths on load. This
change adds a new method to the HDVizData to return that value
as getExtremaWidthsScaled(). An implementation of this scaling
has been added to SimpleHDVizDataImpl. All calls to the original
getExtremaWidths() have been replaced with getExtremaWidthsScaled().

Remaining work needs to be FileCachedHDVizDataImpl to save the
original extremaWidths before scaling and return those original
values in the original getExtremaWidths() method.