Compute and use new scaled layouts in SimpleHDVizDataImpl
In order to write passing tests for comparing the equivalence
between FileCachedHDVizDataImpl and SimpleHDVizDataImpl, some
linear algebra scaling operations needed to be applied to the
layout data. For now, these scaled layouts have been added to
SimpleHDVizDataImpl as a computational step on initialization.
These scaled matrices are returned instead in the getLayout
method. At some ponit, this computation should be factored out
into the processor or some other component.