Client commands now uses promises.
Client commands now return promises to the caller. The promise
is resolved when the server send back a message matching the
corresponding messageId of the outgoing message.

Currently, this only supports single message communications. This
will need to be extended to support multiple-responses for some
of the more complex datasets of large binary data.